pyproject.toml
[tool.poetry]
name = "ogame-professor"
version = "0.1.0"
description = "Web-App for giving calculation services for players in Ogame."
authors = ["kalume <abrahammoracejas@gmail.com>"]
license = "MIT"
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"


psycopg = "^3.2.1"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"
[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
manage = "manage:main"